be ref() manage DTO, sanitize query values
