--- be/src/dto/refresh-token.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class RefreshTokenDto {
-  @IsString()
-  @IsNotEmpty()
-  refresh_token: string;
-}
